+++ src/test/java/chess/domain/board/InitialBoardTest.java
@@ -1,9 +1,7 @@
-package domain.board;
+package chess.domain.board;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import chess.domain.board.InitialBoard;
-import chess.domain.board.Position;
 import chess.domain.piece.Bishop;
 import chess.domain.piece.King;
 import chess.domain.piece.Knight;
