+++ src/test/java/chess/domain/piece/KnightTest.java
@@ -1,51 +1,47 @@
-package domain.piece;
+package chess.domain.piece;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
 import chess.domain.board.Position;
-import chess.domain.piece.Bishop;
-import chess.domain.piece.Blank;
-import chess.domain.piece.Piece;
-import chess.domain.piece.Team;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 
-class BishopTest {
+public class KnightTest {
 
 	@Test
 	void checkBlackTeamSymbol() {
-		Piece bishop = new Bishop(Team.BLACK);
-		assertThat(bishop.getSymbol()).isEqualTo("B");
+		Piece knight = new Knight(Team.BLACK);
+		assertThat(knight.getSymbol()).isEqualTo("N");
 	}
 
 	@Test
 	void checkWhiteTeamSymbol() {
-		Piece bishop = new Bishop(Team.WHITE);
-		assertThat(bishop.getSymbol()).isEqualTo("b");
+		Piece knight = new Knight(Team.WHITE);
+		assertThat(knight.getSymbol()).isEqualTo("n");
 	}
 
 	@ParameterizedTest(name = "[{index}] - to {0}, {1}")
-	@CsvSource(value = {"8, 8", "3, 5", "3, 3", "5, 3"})
+	@CsvSource(value = {"2, 5", "2, 3", "6, 5", "6, 3", "5, 6", "3, 6", "5, 2", "3, 2"})
 	void validateMovement(int targetRow, int targetCol) {
 		Position source = Position.of(4, 4);
-		Piece sourceBishop = new Bishop(Team.BLACK);
+		Piece sourceKnight = new Knight(Team.BLACK);
 		Position target = Position.of(targetRow, targetCol);
-		Piece targetBishop = new Bishop(Team.WHITE);
+		Piece targetKnight = new Knight(Team.WHITE);
 
-		assertDoesNotThrow(() -> sourceBishop.validateMovement(source, target, targetBishop));
+		assertDoesNotThrow(() -> sourceKnight.validateMovement(source, target, targetKnight));
 	}
 
 	@Test
 	void validateDirectionException() {
 		Position source = Position.of(1, 1);
-		Piece bishop = new Bishop(Team.BLACK);
-		Position target = Position.of(1, 2);
+		Piece knight = new Knight(Team.BLACK);
+		Position target = Position.of(2, 2);
 		Piece blank = new Blank();
 
-		assertThatThrownBy(() -> bishop.validateMovement(source, target, blank))
+		assertThatThrownBy(() -> knight.validateMovement(source, target, blank))
 				.isInstanceOf(IllegalArgumentException.class)
 				.hasMessageContaining("해당 기물은 그곳으로 이동할 수 없습니다.");
 	}
@@ -53,11 +49,11 @@ class BishopTest {
 	@Test
 	void validateCatchAllyException() {
 		Position source = Position.of(4, 4);
-		Piece sourceBishop = new Bishop(Team.BLACK);
-		Position target = Position.of(3, 5);
-		Piece targetBishop = new Bishop(Team.BLACK);
+		Piece sourceKnight = new Knight(Team.BLACK);
+		Position target = Position.of(2, 5);
+		Piece targetKnight = new Knight(Team.BLACK);
 
-		assertThatThrownBy(() -> sourceBishop.validateMovement(source, target, targetBishop))
+		assertThatThrownBy(() -> sourceKnight.validateMovement(source, target, targetKnight))
 				.isInstanceOf(IllegalArgumentException.class)
 				.hasMessageContaining("같은 팀의 기물을 잡을 수 없습니다.");
 	}
