+++ src/test/java/chess/domain/state/EndGameTest.java
@@ -1,31 +1,28 @@
-package domain.state;
+package chess.domain.state;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
-import chess.domain.piece.King;
 import chess.domain.piece.Pawn;
 import chess.domain.piece.Team;
-import chess.domain.state.State;
-import chess.domain.state.WhiteTurn;
 import org.junit.jupiter.api.Test;
 
-class KingDeathTest {
+class EndGameTest {
 
 	@Test
 	void isFinished() {
 		State state = new WhiteTurn();
-		State kingDeath = state.play(new Pawn(Team.WHITE), new King(Team.BLACK));
+		State endGame = state.finish();
 
-		assertThat(kingDeath.isFinished()).isTrue();
+		assertThat(endGame.isFinished()).isTrue();
 	}
 
 	@Test
 	void play() {
 		State state = new WhiteTurn();
-		State kingDeath = state.play(new Pawn(Team.WHITE), new King(Team.BLACK));
+		State endGame = state.finish();
 
-		assertThatThrownBy(() -> kingDeath.play(new Pawn(Team.BLACK), new Pawn(Team.WHITE)))
+		assertThatThrownBy(() -> endGame.play(new Pawn(Team.BLACK), new Pawn(Team.WHITE)))
 				.isInstanceOf(IllegalArgumentException.class)
 				.hasMessageContaining("현재는 게임을 진행할 수 없습니다.");
 	}
@@ -33,9 +30,9 @@ class KingDeathTest {
 	@Test
 	void finish() {
 		State state = new WhiteTurn();
-		State kingDeath = state.play(new Pawn(Team.WHITE), new King(Team.BLACK));
+		State endGame = state.finish();
 
-		assertThatThrownBy(kingDeath::finish)
+		assertThatThrownBy(endGame::finish)
 				.isInstanceOf(IllegalArgumentException.class)
 				.hasMessageContaining("게임이 이미 종료되었습니다.");
 	}
